Add JetBrains editor files that are actually useful for those using any Jetbrains editor.
All editor files must be sure to be relative to the project directory and contain no
machine-identifiable information in them such as where the project lives or
where any of your dependencies live.

Other editor files will be accepted to the extent they allow for the above and are good at
giving others a consistent experience.